gaussian_model.py

+++ scene/gaussian_model.py
@@ -166,10 +166,10 @@ class GaussianModel:
 
         opacities = self.inverse_opacity_activation(0.1 * torch.ones((fused_point_cloud.shape[0], 1), dtype=torch.float, device="cuda"))
 
-        self._xyz = nn.Parameter(fused_point_cloud.half().requires_grad_(True))  # float16
-        self._scaling = nn.Parameter(scales.half().requires_grad_(True))  # float16
-        self._features_dc = nn.Parameter(features[:, :, 0:1].transpose(1, 2).contiguous().half().requires_grad_(True))  # float16
-        self._features_rest = nn.Parameter(features[:, :, 1:].transpose(1, 2).contiguous().half().requires_grad_(True))  # float16
+        self._xyz = nn.Parameter(fused_point_cloud.float().requires_grad_(True))  # float32
+        self._scaling = nn.Parameter(scales.float().requires_grad_(True))  # float32
+        self._features_dc = nn.Parameter(features[:, :, 0:1].transpose(1, 2).contiguous().float().requires_grad_(True))  # float32
+        self._features_rest = nn.Parameter(features[:, :, 1:].transpose(1, 2).contiguous().float().requires_grad_(True))  # float32
         self._scaling = nn.Parameter(scales.requires_grad_(True))
         self._rotation = nn.Parameter(rots.requires_grad_(True))
         self._opacity = nn.Parameter(opacities.requires_grad_(True))
